+++ epub3itizer/__init__.py
@@ -1,3 +1,4 @@
 from .conversion import convert_epub2_to_epub3
+from .repair import repair_epub
 
-__all__ = ["convert_epub2_to_epub3"]
+__all__ = ["convert_epub2_to_epub3", "repair_epub"]